Refactor: remove local get_anti_affinity and parsing function definitions from the original file; update imports to use get_anti_affinity from affinity_utils.py and parse_prolog, parse_paths, parse_paths_no_reliability, and parse_allocation from prolog_parse_utils.py
